improved datagenerator + renamed
